Promtail: Teach it to read logs from K8s API server
Normally promtail tails Pod logs by mounting the host path. But in
locked down Kubernetes environments this host path mount may not
be allowed. This may be especially typical in multi-tenant setups.
However, Kubernetes exposes a REST API to fetch logs from.

This change hacks up the existing promtail code to fetch logs
from the API server. It assumes promtail itself is running as a
Pod.

Issues with this change:
- Logging, Code in general need cleanup